websummarizer_crew/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.5.1+cpu
crewai[tools]>=0.86.0,<1.0.0
stripe>=11.0.0
python-dotenv>=1.0.0
requests>=2.31.0
langchain>=0.2.2
langchain-community>=0.0.24
langchain-huggingface>=0.1.2